% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sta344.R
\docType{data}
\name{avg_temp}
\alias{avg_temp}
\title{Average weekly temperatures in Durham, NC}
\format{
A data frame with 156 rows and 3 columns:
\itemize{
\item \code{date} - date representing each week (Wednesday)
\item \code{avg_temp} - average weekly temperature in degrees Farhenheight
\item \code{week} - numerical index of week from the start of the data (0-based)
}
}
\usage{
avg_temp
}
\description{
Data collected from the Darksky weather API and aggregated to weekly averages.
}
\keyword{datasets}
