% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crps.R
\name{calc_crps}
\alias{calc_crps}
\title{Calculate CRPS}
\usage{
calc_crps(x, obs)
}
\arguments{
\item{x}{Samples for the predictive distribution}

\item{obs}{True observed value}
}
\description{
Calculates the the continuous rank probability score
}
