% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{cond_predict}
\alias{cond_predict}
\title{GP conditional predictions}
\usage{
cond_predict(x_pred, x, cov, ..., reps = 1000, diag_adj = 1e-06)
}
\arguments{
\item{x_pred}{Vector or matrix of prediction locations (values).}

\item{x}{Vector or matrix of fitted locations (values).}

\item{cov}{Function. Covariance function that takes a distance matrix as input.}

\item{...}{Parameter values used by \code{cov}.}

\item{reps}{Numeric. Number of samples to draw from conditional distribution.}

\item{diag_adj}{Numeric. Small value that is added covariance diagonal to avoid
numerical singularity issues.}
}
\value{
A matrix of size \code{n_pred} x \code{reps}.
}
\description{
Given a covariance function and parameters this samples from
the conditional MVN of y_pred | y
}
