% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance.R
\name{covariance_functions}
\alias{covariance_functions}
\alias{nugget_cov}
\alias{exp_cov}
\alias{sq_exp_cov}
\alias{pow_exp_cov}
\alias{rquad_cov}
\alias{periodic_cov}
\alias{matern_cov}
\alias{sphere_cov}
\alias{linear_cov}
\title{Covariance functions}
\usage{
nugget_cov(d, sigma2 = 1)

exp_cov(d, sigma2 = 1, l = 1, sigma2_w = 0)

sq_exp_cov(d, sigma2 = 1, l = 1, sigma2_w = 0)

pow_exp_cov(d, sigma2 = 1, l = 1, p = 2)

rquad_cov(d, sigma2 = 1, l = 1, a = 1)

periodic_cov(d, sigma2 = 1, l = 1, per = 1)

matern_cov(d, sigma2 = 1, l = 1, nu = 1/2)

sphere_cov(d, sigma2 = 1, l = 1)

linear_cov(x1, x2, sigma2_b = 1, sigma2_v = 1, c = 0)
}
\arguments{
\item{d}{Distance matrix}

\item{sigma2}{Scale variance parameter}

\item{l}{Inverse length scale parameter}

\item{sigma2_w}{Nugget variance parameter}

\item{p}{Power parameter (powered exponential covariance)}

\item{a}{Alpha parameter (rational quartic covariance)}

\item{per}{Period parameter (periodic covariance)}

\item{nu}{Nu parameter (Matern covariance)}
}
\description{
Functions for calculating various covariances
}
