% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_pred_draws.R
\name{draws_fix}
\alias{draws_fix}
\alias{predicted_draws_fix}
\alias{epred_draws_fix}
\alias{residual_draws_fix}
\title{Stopgap fixes for tidybayes draws functions}
\usage{
predicted_draws_fix(object, newdata, ...)

epred_draws_fix(object, newdata, ...)

residual_draws_fix(object, newdata, ...)
}
\arguments{
\item{object}{A supported Bayesian model fit that can provide fits and predictions.}

\item{newdata}{Data frame to generate predictions from.}

\item{...}{Any additional arguments to be passed to the base function.}
}
\description{
Wrappers around tidybayes draw functions that maintain the
\code{.iteration}, \code{.draw}, and \code{.chain} columns.
}
\seealso{
\code{\link[tidybayes:add_predicted_draws]{tidybayes::predicted_draws()}}
}
