% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semivariogram.R
\name{emp_semivariogram}
\alias{emp_semivariogram}
\title{Estimate the empirical semivariogram}
\usage{
emp_semivariogram(d, y, x, bin = FALSE, binwidth, range_max = NULL)
}
\arguments{
\item{d}{Data frame containing data}

\item{y}{Unquoted variable name for y values}

\item{x}{Unquoted variable name for x values}

\item{bin}{Logical. Should variance estimatess be binned.}

\item{binwidth}{Numeric. The width of the bins to use.}

\item{range_max}{Numeric. Filter bins with a midpoint above this value.
Defaults to 1/3 of the max distance between \code{x} values.}
}
\description{
Calculates the empirical semivariogram for a given
binwidth.
}
