% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{rmvnorm}
\alias{rmvnorm}
\title{Sample from MVN normal}
\usage{
rmvnorm(n, mu = rep(0, nrow(Sigma)), Sigma, diag_adj = 1e-06)
}
\arguments{
\item{n}{Number of draws to sample}

\item{mu}{Mean vector}

\item{Sigma}{Covariance matrix}

\item{diag_adj}{Numeric. Small value that is added covariance's diagonal to avoid
numerical singularity issues.}
}
\value{
A matrix of size nrow(Sigma) x n
}
\description{
Generate draws of the given MVN distribution.
}
