% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semivariogram.R
\name{sv_functions}
\alias{sv_functions}
\alias{exp_sv}
\alias{sq_exp_sv}
\alias{pow_exp_sv}
\title{Semivariogram functions}
\usage{
exp_sv(d, sigma2 = 1, l = 1, sigma2_w = 0)

sq_exp_sv(d, sigma2 = 1, l = 1, sigma2_w = 0)

pow_exp_sv(d, sigma2 = 1, l = 1, p = 2, sigma2_w = 0)
}
\arguments{
\item{d}{Distance matrix}

\item{sigma2}{Scale variance parameter}

\item{l}{Inverse length scale parameter}

\item{sigma2_w}{Nugget variance parameter}

\item{p}{Power parameter (powered exponential covariance)}
}
\description{
Functions for calculating various semivariograms
}
